Create README.md
I am a Systems Engineer graduated in 1999, graduated with an academic diploma in 2000 and a degree in national provision processed in 2004.

I have a diploma in information and communication technology in 2002, a master's degree in intelligent flexible systems (Artificial Intelligence) in 2003. I have complemented my studies with a diploma in higher education in 2007.

I have taken several refresher courses and training both face-to-face and virtual throughout my 20 years of professional life.

I have worked as an ORACLE DBA, also as an administrator of Unix and Linux operating systems, for some time I have specialized in JAVA programming with 15 years of experience in this language. Currently and for 5 years, I have specialized in application development using next-generation languages such as JavaScript, mainly oriented to the creation of applications through NodeJS for backend and JavaScript for frontend, with the use of some Frameworks such as Angular, VUE or REACT.

I have been a software developer consultant, IT external auditor with more than 10 public and private companies audited in the IT area and I have provided IT support in several national and international institutions.

I have worked for several semesters as a professor at different universities, I have been invited to advise thesis for undergraduate degrees and I have also represented the Society of Engineers of Bolivia on several occasions in the final defenses of Degree Projects and Thesis.

I have created and developed virtual courses at a technical level for INFOCAL, having been published through the newspaper La Razón several years ago.

I am the manager and founder of the software and intelligent solutions development company, BearSoft, with several projects developed for different national and international companies.

Currently creator of the SmartDecisions platform, developed in JavaScript and NodeJS working to provide REST API services for companies in the United States (EFFEX and BOXTRIBE). In Bolivia I also have a platform for system integration of AGETIC with supermarkets (FIDALGA and KETAL) for payment of subsidies and double bonus. More information about me and BearSoft, can be seen in my web portal http://bearsoft.com.bo.